requirements.txt
# Core dependencies
openai>=1.3.0
anthropic>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0

# Model serving dependencies
transformers>=4.35.0
torch>=2.0.0
pillow>=10.0.0

# Dataset dependencies
datasets>=2.14.0
tqdm>=4.64.0

# Testing dependencies
pytest>=7.0.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Workflow orchestration
networkx>=3.0
pyyaml>=6.0

# Vector database
chromadb>=0.4.0
pypdf>=3.0.0

# Optional dependencies for future features
# pydantic>=2.0.0  # For data validation
# datasets>=2.14.0  # For benchmarking datasets
# fastapi>=0.104.0  # For potential web UI
# uvicorn>=0.24.0  # For FastAPI server
# psutil>=5.9.0  # For resource monitoring